Control.Monad.Class.MonadAsync

Documentation

class (MonadSTM m, MonadThread m, MonadAsyncSTM (Async m) (STM m)) ⇒ MonadAsync m where

Minimal complete definition

async, asyncThreadId, cancel, cancelWith

Associated Types

type Async m ∷ ★ → ★

An asynchronous action

Methods

async ∷ m a → m (Async m a)

asyncThreadId ∷ Proxy m → Async m a → ThreadId m

withAsync ∷ m a → (Async m a → m b) → m b

wait ∷ Async m a → m a

poll ∷ Async m a → m (Maybe (Either SomeException a))

waitCatch ∷ Async m a → m (Either SomeException a)

cancel ∷ Async m a → m ()

cancelWith ∷ Exception e ⇒ Async m a → e → m ()

uninterruptibleCancel ∷ Async m a → m ()

waitAny ∷ [Async m a] → m (Async m a, a)

waitAnyCatch ∷ [Async m a] → m (Async m a, Either SomeException a)

waitAnyCancel ∷ [Async m a] → m (Async m a, a)

waitAnyCatchCancel ∷ [Async m a] → m (Async m a, Either SomeException a)

waitEither ∷ Async m a → Async m b → m (Either a b)

waitEitherCatch ∷ Async m a → Async m b → m (Either (Either SomeException a) (Either SomeException b))

Note, IO-based implementations should override the default implementation. See the async package implementation and comments. http://hackage.haskell.org/package/async-2.2.1/docs/src/Control.Concurrent.Async.html#waitEitherCatch

waitEitherCancel ∷ Async m a → Async m b → m (Either a b)

waitEitherCatchCancel ∷ Async m a → Async m b → m (Either (Either SomeException a) (Either SomeException b))

waitEither_ ∷ Async m a → Async m b → m ()

waitBoth ∷ Async m a → Async m b → m (a, b)

race ∷ m a → m b → m (Either a b)

race_ ∷ m a → m b → m ()

concurrently ∷ m a → m b → m (a, b)

withAsync ∷ MonadMask m ⇒ m a → (Async m a → m b) → m b

uninterruptibleCancel ∷ MonadMask m ⇒ Async m a → m ()

waitAnyCancel ∷ MonadThrow m ⇒ [Async m a] → m (Async m a, a)

waitAnyCatchCancel ∷ MonadThrow m ⇒ [Async m a] → m (Async m a, Either SomeException a)

waitEitherCancel ∷ MonadThrow m ⇒ Async m a → Async m b → m (Either a b)

waitEitherCatchCancel ∷ MonadThrow m ⇒ Async m a → Async m b → m (Either (Either SomeException a) (Either SomeException b))

class (Functor async, MonadSTMTx stm) ⇒ MonadAsyncSTM async stm where

Minimal complete definition

waitCatchSTM, pollSTM

Methods

waitSTM ∷ async a → stm a

pollSTM ∷ async a → stm (Maybe (Either SomeException a))

waitCatchSTM ∷ async a → stm (Either SomeException a)

waitSTM ∷ MonadThrow stm ⇒ async a → stm a

waitAnySTM ∷ [async a] → stm (async a, a)

waitAnyCatchSTM ∷ [async a] → stm (async a, Either SomeException a)

waitEitherSTM ∷ async a → async b → stm (Either a b)

waitEitherSTM_ ∷ async a → async b → stm ()

waitEitherCatchSTM ∷ async a → async b → stm (Either (Either SomeException a) (Either SomeException b))

waitBothSTM ∷ async a → async b → stm (a, b)

waitAnySTM ∷ MonadThrow stm ⇒ [async a] → stm (async a, a)

waitEitherSTM ∷ MonadThrow stm ⇒ async a → async b → stm (Either a b)

waitEitherSTM_ ∷ MonadThrow stm ⇒ async a → async b → stm ()

waitBothSTM ∷ MonadThrow stm ⇒ async a → async b → stm (a, b)

data AsyncCancelled

The exception thrown by cancel to terminate a thread.

Constructors

AsyncCancelled

data ExceptionInLinkedThread

Exception from child thread re-raised in parent thread

We record the thread ID of the child thread as a String. This avoids an m parameter in the type, which is important: ExceptionInLinkedThread must be an instance of Exception, requiring it to be Typeable; if m appeared in the type, we would require m to be Typeable, which does not work with with the simulator, as it would require a Typeable constraint on the s parameter of SimM.

Constructors

ExceptionInLinkedThread String SomeException

link ∷ (MonadAsync m, MonadFork m, MonadMask m) ⇒ Async m a → m ()

linkTo ∷ (MonadAsync m, MonadFork m, MonadMask m) ⇒ ThreadId m → Async m a → m ()

Generalizion of link that links an async to an arbitrary thread.

linkOnly ∷ ∀ m a. (MonadAsync m, MonadFork m, MonadMask m) ⇒ (SomeException → Bool) → Async m a → m ()

linkToOnly ∷ ∀ m a. (MonadAsync m, MonadFork m, MonadMask m) ⇒ ThreadId m → (SomeException → Bool) → Async m a → m ()