Control.Monad.Class.MonadSTM.Strict

Documentation

data TBQueueDefault m a

Constructors

TBQueue !(TVar m Natural) !(TVar m [a]) !(TVar m Natural) !(TVar m [a]) !Natural

data TQueueDefault m a

Constructors

TQueue !(TVar m [a]) !(TVar m [a])

data TMVarDefault m a

class (MonadSTM m, MonadLabelledSTMTx (STM m)) ⇒ MonadLabelledSTM m where

A convenience class which provides MonadSTM and MonadLabelledSTMTx constraints.

Minimal complete definition

Nothing

Methods

labelTQueueIO ∷ TQueue m a → String → m ()

default labelTQueueIO ∷ TQueue m a → String → m ()

labelTBQueueIO ∷ TBQueue m a → String → m ()

default labelTBQueueIO ∷ TBQueue m a → String → m ()

class MonadSTMTx stm ⇒ MonadLabelledSTMTx stm where

Labelled TVars, TMVars, TQueues and TBQueues.

Minimal complete definition

labelTVar, labelTMVar, labelTQueue, labelTBQueue

Methods

labelTQueue ∷ TQueue_ stm a → String → stm ()

labelTBQueue ∷ TBQueue_ stm a → String → stm ()

class (Monad m, MonadSTMTx (STM m)) ⇒ MonadSTM m where

Minimal complete definition

atomically

Associated Types

type STM m ∷ Type → Type

Methods

atomically ∷ HasCallStack ⇒ STM m a → m a

newTQueueIO ∷ m (TQueue m a)

newTBQueueIO ∷ Natural → m (TBQueue m a)

type TBQueue m = TBQueue_ (STM m)

type TQueue m = TQueue_ (STM m)

class (Monad stm, Alternative stm, MonadPlus stm) ⇒ MonadSTMTx stm where

Minimal complete definition

newTVar, readTVar, writeTVar, retry, orElse, newTMVar, newEmptyTMVar, takeTMVar, tryTakeTMVar, putTMVar, tryPutTMVar, readTMVar, tryReadTMVar, swapTMVar, isEmptyTMVar, newTQueue, readTQueue, tryReadTQueue, peekTQueue, tryPeekTQueue, writeTQueue, isEmptyTQueue, newTBQueue, readTBQueue, tryReadTBQueue, peekTBQueue, tryPeekTBQueue, flushTBQueue, writeTBQueue, lengthTBQueue, isEmptyTBQueue, isFullTBQueue

Associated Types

type TVar_ stm ∷ Type → Type

type TMVar_ stm ∷ Type → Type

type TQueue_ stm ∷ Type → Type

type TBQueue_ stm ∷ Type → Type

Methods

retry ∷ stm a

orElse ∷ stm a → stm a → stm a

modifyTVar' ∷ TVar_ stm a → (a → a) → stm ()

check ∷ Bool → stm ()

newTQueue ∷ stm (TQueue_ stm a)

readTQueue ∷ TQueue_ stm a → stm a

tryReadTQueue ∷ TQueue_ stm a → stm (Maybe a)

peekTQueue ∷ TQueue_ stm a → stm a

tryPeekTQueue ∷ TQueue_ stm a → stm (Maybe a)

writeTQueue ∷ TQueue_ stm a → a → stm ()

isEmptyTQueue ∷ TQueue_ stm a → stm Bool

newTBQueue ∷ Natural → stm (TBQueue_ stm a)

readTBQueue ∷ TBQueue_ stm a → stm a

tryReadTBQueue ∷ TBQueue_ stm a → stm (Maybe a)

peekTBQueue ∷ TBQueue_ stm a → stm a

tryPeekTBQueue ∷ TBQueue_ stm a → stm (Maybe a)

flushTBQueue ∷ TBQueue_ stm a → stm [a]

writeTBQueue ∷ TBQueue_ stm a → a → stm ()

lengthTBQueue ∷ TBQueue_ stm a → stm Natural

Since: 0.2.0.0

isEmptyTBQueue ∷ TBQueue_ stm a → stm Bool

isFullTBQueue ∷ TBQueue_ stm a → stm Bool

labelTMVarDefault ∷ MonadLabelledSTM m ⇒ TMVarDefault m a → String → STM m ()

newTMVarDefault ∷ MonadSTM m ⇒ a → STM m (TMVarDefault m a)

newTMVarIODefault ∷ MonadSTM m ⇒ a → m (TMVarDefault m a)

newTMVarMDefault ∷ MonadSTM m ⇒ a → m (TMVarDefault m a)

Deprecated: Use newTMVarIODefault

newEmptyTMVarDefault ∷ MonadSTM m ⇒ STM m (TMVarDefault m a)

newEmptyTMVarIODefault ∷ MonadSTM m ⇒ m (TMVarDefault m a)

newEmptyTMVarMDefault ∷ MonadSTM m ⇒ m (TMVarDefault m a)

Deprecated: Use newEmptyTMVarIODefault

takeTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → STM m a

tryTakeTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → STM m (Maybe a)

putTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → a → STM m ()

tryPutTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → a → STM m Bool

readTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → STM m a

tryReadTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → STM m (Maybe a)

swapTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → a → STM m a

isEmptyTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → STM m Bool

labelTQueueDefault ∷ MonadLabelledSTM m ⇒ TQueueDefault m a → String → STM m ()

newTQueueDefault ∷ MonadSTM m ⇒ STM m (TQueueDefault m a)

writeTQueueDefault ∷ MonadSTM m ⇒ TQueueDefault m a → a → STM m ()

readTQueueDefault ∷ MonadSTM m ⇒ TQueueDefault m a → STM m a

tryReadTQueueDefault ∷ MonadSTMTx (STM m) ⇒ TQueueDefault m a → STM m (Maybe a)

isEmptyTQueueDefault ∷ MonadSTMTx (STM m) ⇒ TQueueDefault m a → STM m Bool

peekTQueueDefault ∷ MonadSTM m ⇒ TQueueDefault m a → STM m a

tryPeekTQueueDefault ∷ MonadSTM m ⇒ TQueueDefault m a → STM m (Maybe a)

labelTBQueueDefault ∷ MonadLabelledSTM m ⇒ TBQueueDefault m a → String → STM m ()

newTBQueueDefault ∷ MonadSTM m ⇒ Natural → STM m (TBQueueDefault m a)

readTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m a

tryReadTBQueueDefault ∷ MonadSTMTx (STM m) ⇒ TBQueueDefault m a → STM m (Maybe a)

peekTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m a

tryPeekTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m (Maybe a)

writeTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → a → STM m ()

isEmptyTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m Bool

isFullTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m Bool

lengthTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m Natural

flushTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m [a]

throwSTM ∷ (MonadSTMTx stm, MonadThrow stm, Exception e) ⇒ e → stm a

throwIO specialised to stm monad.

catchSTM ∷ (MonadSTMTx stm, MonadCatch stm, Exception e) ⇒ stm a → (e → stm a) → stm a

catch speclialized for an stm monad.

type LazyTVar m = TVar m

type LazyTMVar m = TMVar m

StrictTVar

data StrictTVar m a

labelTVar ∷ MonadLabelledSTM m ⇒ StrictTVar m a → String → STM m ()

labelTVarIO ∷ MonadLabelledSTM m ⇒ StrictTVar m a → String → m ()

castStrictTVar ∷ LazyTVar m ~ LazyTVar n ⇒ StrictTVar m a → StrictTVar n a

toLazyTVar ∷ StrictTVar m a → LazyTVar m a

Get the underlying TVar

Since we obviously cannot guarantee that updates to this LazyTVar will be strict, this should be used with caution.

newTVar ∷ MonadSTM m ⇒ a → STM m (StrictTVar m a)

newTVarIO ∷ MonadSTM m ⇒ a → m (StrictTVar m a)

newTVarWithInvariantIO

Arguments

∷ (MonadSTM m, HasCallStack)

⇒ (a → Maybe String)
Invariant (expect Nothing)

→ a

→ m (StrictTVar m a)

readTVar ∷ MonadSTM m ⇒ StrictTVar m a → STM m a

readTVarIO ∷ MonadSTM m ⇒ StrictTVar m a → m a

writeTVar ∷ (MonadSTM m, HasCallStack) ⇒ StrictTVar m a → a → STM m ()

modifyTVar ∷ MonadSTM m ⇒ StrictTVar m a → (a → a) → STM m ()

stateTVar ∷ MonadSTM m ⇒ StrictTVar m s → (s → (a, s)) → STM m a

swapTVar ∷ MonadSTM m ⇒ StrictTVar m a → a → STM m a

StrictTMVar

data StrictTMVar m a

labelTMVar ∷ MonadLabelledSTM m ⇒ StrictTMVar m a → String → STM m ()

labelTMVarIO ∷ MonadLabelledSTM m ⇒ StrictTMVar m a → String → m ()

castStrictTMVar ∷ LazyTMVar m ~ LazyTMVar n ⇒ StrictTMVar m a → StrictTMVar n a

newTMVar ∷ MonadSTM m ⇒ a → STM m (StrictTMVar m a)

newTMVarIO ∷ MonadSTM m ⇒ a → m (StrictTMVar m a)

newEmptyTMVar ∷ MonadSTM m ⇒ STM m (StrictTMVar m a)

newEmptyTMVarIO ∷ MonadSTM m ⇒ m (StrictTMVar m a)

takeTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → STM m a

tryTakeTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → STM m (Maybe a)

putTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → a → STM m ()

tryPutTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → a → STM m Bool

readTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → STM m a

tryReadTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → STM m (Maybe a)

swapTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → a → STM m a

isEmptyTMVar ∷ MonadSTM m ⇒ StrictTMVar m a → STM m Bool

Low-level API

checkInvariant ∷ HasCallStack ⇒ Maybe String → a → a

Check invariant (if enabled) before continuing

checkInvariant mErr x is equal to x if mErr == Nothing, and throws an error err if mErr == Just err.

This is exported so that other code that wants to conditionally check invariants can reuse the same logic, rather than having to introduce new per-package flags.

Deprecated API

updateTVar ∷ MonadSTM m ⇒ StrictTVar m s → (s → (a, s)) → STM m a

Deprecated: Use stateTVar

newTVarM ∷ MonadSTM m ⇒ a → m (StrictTVar m a)

Deprecated: Use newTVarIO

newTVarWithInvariantM

Arguments

∷ (MonadSTM m, HasCallStack)

⇒ (a → Maybe String)
Invariant (expect Nothing)

→ a

→ m (StrictTVar m a)

Deprecated: Use newTVarWithInvariantIO

newTMVarM ∷ MonadSTM m ⇒ a → m (StrictTMVar m a)

Deprecated: Use newTVarIO

newEmptyTMVarM ∷ MonadSTM m ⇒ m (StrictTMVar m a)

Deprecated: Use newEmptyTMVarIO