Control.Monad.Class.MonadThrow

Documentation

class Monad m ⇒ MonadThrow m where

Throwing exceptions, and resource handling in the presence of exceptions.

Does not include the ability to respond to exceptions.

Minimal complete definition

throwM

Methods

throwM ∷ Exception e ⇒ e → m a

bracket ∷ m a → (a → m b) → (a → m c) → m c

bracket_ ∷ m a → m b → m c → m c

finally ∷ m a → m b → m a

bracket ∷ MonadCatch m ⇒ m a → (a → m b) → (a → m c) → m c

class MonadThrow m ⇒ MonadCatch m where

Catching exceptions.

Covers standard utilities to respond to exceptions.

Minimal complete definition

catch

Methods

catch ∷ Exception e ⇒ m a → (e → m a) → m a

catchJust ∷ Exception e ⇒ (e → Maybe b) → m a → (b → m a) → m a

try ∷ Exception e ⇒ m a → m (Either e a)

tryJust ∷ Exception e ⇒ (e → Maybe b) → m a → m (Either b a)

handle ∷ Exception e ⇒ (e → m a) → m a → m a

handleJust ∷ Exception e ⇒ (e → Maybe b) → (b → m a) → m a → m a

onException ∷ m a → m b → m a

bracketOnError ∷ m a → (a → m b) → (a → m c) → m c

generalBracket ∷ m a → (a → ExitCase b → m c) → (a → m b) → m (b, c)

General form of bracket

See http://hackage.haskell.org/package/exceptions-0.10.0/docs/Control-Monad-Catch.html#v:generalBracket for discussion and motivation.

generalBracket ∷ MonadMask m ⇒ m a → (a → ExitCase b → m c) → (a → m b) → m (b, c)

General form of bracket

See http://hackage.haskell.org/package/exceptions-0.10.0/docs/Control-Monad-Catch.html#v:generalBracket for discussion and motivation.

class MonadCatch m ⇒ MonadMask m where

Support for safely working in the presence of asynchronous exceptions.

This is typically not needed directly as the utilities in MonadThrow and MonadCatch cover most use cases.

Minimal complete definition

mask, uninterruptibleMask

Methods

mask ∷ ((∀ a. m a → m a) → m b) → m b

uninterruptibleMask ∷ ((∀ a. m a → m a) → m b) → m b

mask_ ∷ m a → m a

uninterruptibleMask_ ∷ m a → m a

class (Typeable e, Show e) ⇒ Exception e where

Any type that you wish to throw or catch as an exception must be an instance of the Exception class. The simplest case is a new exception type directly below the root:

data MyException = ThisException | ThatException
    deriving Show

instance Exception MyException

The default method definitions in the Exception class do what we need in this case. You can now throw and catch ThisException and ThatException as exceptions:

*Main> throw ThisException `catch` \e -> putStrLn ("Caught " ++ show (e :: MyException))
Caught ThisException

In more complicated examples, you may wish to define a whole hierarchy of exceptions:

---------------------------------------------------------------------
-- Make the root exception type for all the exceptions in a compiler

data SomeCompilerException = forall e . Exception e => SomeCompilerException e

instance Show SomeCompilerException where
    show (SomeCompilerException e) = show e

instance Exception SomeCompilerException

compilerExceptionToException :: Exception e => e -> SomeException
compilerExceptionToException = toException . SomeCompilerException

compilerExceptionFromException :: Exception e => SomeException -> Maybe e
compilerExceptionFromException x = do
    SomeCompilerException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make a subhierarchy for exceptions in the frontend of the compiler

data SomeFrontendException = forall e . Exception e => SomeFrontendException e

instance Show SomeFrontendException where
    show (SomeFrontendException e) = show e

instance Exception SomeFrontendException where
    toException = compilerExceptionToException
    fromException = compilerExceptionFromException

frontendExceptionToException :: Exception e => e -> SomeException
frontendExceptionToException = toException . SomeFrontendException

frontendExceptionFromException :: Exception e => SomeException -> Maybe e
frontendExceptionFromException x = do
    SomeFrontendException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make an exception type for a particular frontend compiler exception

data MismatchedParentheses = MismatchedParentheses
    deriving Show

instance Exception MismatchedParentheses where
    toException   = frontendExceptionToException
    fromException = frontendExceptionFromException

We can now catch a MismatchedParentheses exception as MismatchedParentheses, SomeFrontendException or SomeCompilerException, but not other types, e.g. IOException:

*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: MismatchedParentheses))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeFrontendException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeCompilerException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: IOException))
*** Exception: MismatchedParentheses

Minimal complete definition

Nothing

Methods

toException ∷ e → SomeException

fromException ∷ SomeException → Maybe e

displayException ∷ e → String

Render this exception value in a human-friendly manner.

Default implementation: show.

Since: base-4.8.0.0

data SomeException

The SomeException type is the root of the exception type hierarchy. When an exception of type e is thrown, behind the scenes it is encapsulated in a SomeException.

data ExitCase a

Used in generalBracket

See exceptions package for discussion and motivation.

Constructors

ExitCaseSuccess a

ExitCaseException SomeException

ExitCaseAbort