Control.Monad.Class.MonadTimer

Documentation

class Monad m ⇒ MonadDelay m where

Minimal complete definition

Nothing

Methods

threadDelay ∷ DiffTime → m ()

threadDelay ∷ MonadTimer m ⇒ DiffTime → m ()

class (MonadSTM m, MonadDelay m) ⇒ MonadTimer m where

Minimal complete definition

newTimeout, readTimeout, updateTimeout, cancelTimeout, timeout

Associated Types

data Timeout m ∷ ★

Methods

newTimeout ∷ DiffTime → m (Timeout m)

Create a new timeout which will fire at the given time duration in the future.

The timeout will start in the TimeoutPending state and either fire at or after the given time leaving it in the TimeoutFired state, or it may be cancelled with cancelTimeout, leaving it in the TimeoutCancelled state.

Timeouts cannot be reset to the pending state once fired or cancelled (as this would be very racy). You should create a new timeout if you need this functionality.

readTimeout ∷ Timeout m → STM m TimeoutState

Read the current state of a timeout. This does not block, but returns the current state. It is your responsibility to use retry to wait.

Alternatively you may wish to use the convenience utility awaitTimeout to wait for just the fired or cancelled outcomes.

You should consider the cancelled state if you plan to use cancelTimeout.

updateTimeout ∷ Timeout m → DiffTime → m ()

cancelTimeout ∷ Timeout m → m ()

Cancel a timeout (unless it has already fired), putting it into the TimeoutCancelled state. Code reading and acting on the timeout state need to handle such cancellation appropriately.

It is safe to race this concurrently against the timer firing. It will have no effect if the timer fires first.

awaitTimeout ∷ Timeout m → STM m Bool

Returns True when the timeout is fired, or False if it is cancelled.

registerDelay ∷ DiffTime → m (TVar m Bool)

registerDelay ∷ MonadFork m ⇒ DiffTime → m (TVar m Bool)

timeout ∷ DiffTime → m a → m (Maybe a)

data TimeoutState

Constructors

TimeoutPending

TimeoutFired

TimeoutCancelled